+++ src/main/java/gwt/material/design/amcharts/client/ui/chart/js/options/AmGaugeBand.java
@@ -49,7 +49,7 @@ public class AmGaugeBand {
     public int endValue;
 
     @JsProperty
-    public int[] gradientRatio;
+    public double[] gradientRatio;
 
     @JsProperty
     public String id;
